Split out INSAD classes
